GCP_트위터 스트리밍 데이터 수집/requirements.txt
google-cloud-pubsub==1.6.0
tweepy==3.8.0